image_inspector_app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
Django==5.2.1
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.2
huggingface-hub==0.32.3
idna==3.10
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
